refactor(agents): move subscriber-identity normalization utils to shared/util
email-normalization and phone-normalization are subscriber lookup helpers
consumed by the subscriber resolver and the WhatsApp test-template flow, not
email-specific. Relocate them out of email/util into shared/util.
